pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "ouchi-face-backend"
version = "0.1.0"
description = "Ouchi Face local catalog backend"
readme = "README.md"
authors = [
    { name = "Ouchi Face Team" },
]
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.23",
    "sqlmodel>=0.0.16",
    "apscheduler>=3.10",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "gitpython>=3.1",
    "ruamel.yaml>=0.18",
    "pydantic-settings>=2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools]
package-dir = {"" = "backend"}

[tool.setuptools.packages.find]
where = ["backend"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
